refactor: 코드 가독성 위해 CoreDataError 열거형을 새로운 파일로 옮김
