+++ CloudNotes/CloudNotes/View/Controller/PrimaryViewController.swift
@@ -87,23 +87,3 @@ extension PrimaryViewController {
         }
     }
 }
-
-enum CoreDataError: Error, LocalizedError {
-    case fetchError
-    case saveError
-    case deletError
-    case updateError
-    
-    var errorDescription: String? {
-        switch self {
-        case .fetchError:
-            return "fetch에 실패했습니다."
-        case .saveError:
-            return "save에 실패했습니다."
-        case .deletError:
-            return "delet에 실패했습니다."
-        case .updateError:
-            return "update에 실패했습니다."
-        }
-    }
-}
